# services/api/openapi.json
{
  "components": {
    "schemas": {
...
      "NewWebhook": {
        "properties": {
          "events": {
            "description": "events it subscribes to",
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Events",
            "type": "array"
          },
          "secret": {
            "description": "shared secret to sign; never returned",
            "maxLength": 512,
            "minLength": 16,
            "title": "Secret",
            "type": "string"
          },
          "template": {
            "default": "generic",
            "description": "servicenow, jira, generic…",
            "title": "Template",
            "type": "string"
          },
          "url": {
            "description": "endpoint of the client, https only",
            "maxLength": 2000,
            "pattern": "^https://",
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "events",
          "secret"
        ],
        "title": "NewWebhook",
        "type": "object"
      },
      "Page": {
        "description": "A page of a listing: the items and the cursor to ask for the next one.",
        "properties": {
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Items",
            "type": "array"
          },
          "next": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "cursor of the next page, null at the end",
            "title": "Next"
          }
        },
        "title": "Page",
        "type": "object"
      },
      "Problem": {
        "description": "An error as RFC 9457 describes it: the same shape whatever went wrong.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "what happened in this occurrence",
            "title": "Detail"
          },
          "instance": {
            "description": "path where it happened",
            "title": "Instance",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code",
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "short human summary, the same for every occurrence",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "description": "identifier of the kind of problem",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "instance"
        ],
        "title": "Problem",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "access token issued by the appliance Keycloak",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Campaigns, inventory, findings and evidence of the ARGOS appliance. Listings paginate by opaque cursor, errors follow RFC 9457 and every POST that creates accepts an Idempotency-Key.",
    "title": "ARGOS API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/webhooks": {
      "get": {
        "operationId": "list_all_api_v1_webhooks_get",
        "parameters": [
          {
            "description": "opaque cursor returned by the previous page",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "opaque cursor returned by the previous page",
              "title": "Cursor"
            }
          },
          {
            "description": "items per page",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "items per page",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "no valid token"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "the role does not allow it"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "it does not exist"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "the state does not allow it"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "invalid body"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "not configured yet"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List the subscriptions",
        "tags": [
          "webhooks"
        ]
      },
      "post": {
        "operationId": "subscribe_api_v1_webhooks_post",
        "parameters": [
          {
            "description": "key chosen by the client; repeating it returns the first result",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "key chosen by the client; repeating it returns the first result",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewWebhook"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Subscribe Api V1 Webhooks Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "no valid token"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "the role does not allow it"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "it does not exist"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "the state does not allow it"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "invalid body"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "not configured yet"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Subscribe an endpoint",
        "tags": [
          "webhooks"
        ]
      }
    },
...
  }
}